[SP-5087] [BISERVER-14142] Import Metadata-Pen Server fails to import localization (#1061)
Localization files are imported and saved with a unique ".tmp" name.
The logic to determine what locale a message file is is based on the
file name, and the .tmp name was screwing up the locale determination.

The existing code attempted to store both the "originalFileName"
and the tempFileName, but was setting the values incorrectly.
This change sets the original and tempname correctly, and
stores both in the JSON constructed list of locale bundles.

https://jira.pentaho.com/browse/BISERVER-14142